Added a check for start tour in handleVisitStep.ts
At the start of the tour beforeLeave and afterLeave actions are triggered for the first action together with beforeEnter and afterEnter, this happens because this.activeStep is equal to stepIndex passed to goToStep function. In order to fix that, I'm running equality check for the index before running action for the current step.